// packages/cosmic/src/checkbox/interface.ts
/*
 * Copyright (c) Baidu, Inc. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type Appearance = 'tag' | 'mark';

export interface CheckboxProps {
    appearance: Appearance;
    disabled: boolean;
    value: string | number;
    checked: boolean;
    indeterminate: boolean;
};

export interface CheckboxEvents {
    change: {
        checked: boolean;
        event: Event;
    };
}

export interface CheckboxDataMessages {
    'cos:checkbox-checked-changed': {
        checked: boolean;
        value: string | number;
    };
}

export type CheckboxData = Required<CheckboxProps>;